requirements.txt
streamlit>=1.34.0
requests>=2.31.0
Unidecode>=1.3.8
google-generativeai
